Create find-urls-matching-keyword
This python script accesses xml sitemap of the website and finds all urls that match specific keyword. I use Python  requests library to fetch the sitemap and the xml.etree.ElementTree module to parse the XML. 

Please define the URL of the XML sitemap (sitemap_url) and the keyword to search for (keyword).

The find_urls_with_keyword function retrieves the XML sitemap using the requests.get method and then parses it using xml.etree.ElementTree. It finds all URL elements within the sitemap and checks if the keyword exists in each URL.
URLs containing the keyword are added to the matching_urls list.

The script prints out the URLs that contain the keyword.